Fix broken exceptions AND allow extending client
Description

It wasn't possible to extend the client before, which could be annoying
if a new endpoint was added to Cryptowatch but this client wasn't updated quickly enough.
So instead expose the doApiRequest method so that devs can extend the RestApiClient.

Additionally there were some exceptions in main.dart, one of which was caused
by passing an Iterable to sortPublicOrders which expected a list.